Get required demo functions from defender demo
