+++ FreeRTOS-Plus/Demo/AWS/Fleet_Provisioning_Windows_Simulator/Fleet_Provisioning_With_CSR_Demo/DemoTasks/FleetProvisioningDemoExample.c
@@ -1,0 +1,764 @@
+/*
+ * AWS IoT Device SDK for Embedded C 202108.00
+ * Copyright (C) 2020 Amazon.com, Inc. or its affiliates.  All Rights Reserved.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+/*
+ * Demo for showing use of the Fleet Provisioning library to use the Fleet
+ * Provisioning feature of AWS IoT Core for provisioning devices with
+ * credentials. This demo shows how a device can be provisioned with AWS IoT
+ * Core using the Certificate Signing Request workflow of the Fleet
+ * Provisioning feature.
+ *
+ * The Fleet Provisioning library provides macros and helper functions for
+ * assembling MQTT topics strings, and for determining whether an incoming MQTT
+ * message is related to the Fleet Provisioning API of AWS IoT Core. The Fleet
+ * Provisioning library does not depend on any particular MQTT library,
+ * therefore the functionality for MQTT operations is placed in another file
+ * (mqtt_operations.c). This demo uses the coreMQTT library. If needed,
+ * mqtt_operations.c can be modified to replace coreMQTT with another MQTT
+ * library. This demo requires using the AWS IoT Core broker as Fleet
+ * Provisioning is an AWS IoT Core feature.
+ *
+ * This demo provisions a device certificate using the provisioning by claim
+ * workflow with a Certificate Signing Request (CSR). The demo connects to AWS
+ * IoT Core using provided claim credentials (whose certificate needs to be
+ * registered with IoT Core before running this demo), subscribes to the
+ * CreateCertificateFromCsr topics, and obtains a certificate. It then
+ * subscribes to the RegisterThing topics and activates the certificate and
+ * obtains a Thing using the provisioning template. Finally, it reconnects to
+ * AWS IoT Core using the new credentials.
+ */
+
+/* Standard includes. */
+#include <stdlib.h>
+#include <stdint.h>
+#include <stdbool.h>
+
+/* POSIX includes. */
+
+#include <errno.h>
+
+/* Demo config. */
+#include "demo_config.h"
+
+/* TinyCBOR library for CBOR encoding and decoding operations. */
+#include "cbor.h"
+
+/* corePKCS11 includes. */
+#include "core_pkcs11.h"
+#include "core_pkcs11_config.h"
+
+/* AWS IoT Fleet Provisioning Library. */
+#include "fleet_provisioning.h"
+
+/* Demo includes. */
+#include "mqtt_operations.h"
+#include "pkcs11_operations.h"
+#include "fleet_provisioning_serializer.h"
+
+/**
+ * These configurations are required. Throw compilation error if it is not
+ * defined.
+ */
+#ifndef PROVISIONING_TEMPLATE_NAME
+    #error "Please define PROVISIONING_TEMPLATE_NAME to the template name registered with AWS IoT Core in demo_config.h."
+#endif
+#ifndef CLAIM_CERT_PATH
+    #error "Please define path to claim certificate (CLAIM_CERT_PATH) in demo_config.h."
+#endif
+#ifndef CLAIM_PRIVATE_KEY_PATH
+    #error "Please define path to claim private key (CLAIM_PRIVATE_KEY_PATH) in demo_config.h."
+#endif
+#ifndef DEVICE_SERIAL_NUMBER
+    #error "Please define a serial number (DEVICE_SERIAL_NUMBER) in demo_config.h."
+#endif
+
+/**
+ * @brief The length of #PROVISIONING_TEMPLATE_NAME.
+ */
+#define PROVISIONING_TEMPLATE_NAME_LENGTH    ( ( uint16_t ) ( sizeof( PROVISIONING_TEMPLATE_NAME ) - 1 ) )
+
+/**
+ * @brief The length of #DEVICE_SERIAL_NUMBER.
+ */
+#define DEVICE_SERIAL_NUMBER_LENGTH          ( ( uint16_t ) ( sizeof( DEVICE_SERIAL_NUMBER ) - 1 ) )
+
+/**
+ * @brief Size of AWS IoT Thing name buffer.
+ *
+ * See https://docs.aws.amazon.com/iot/latest/apireference/API_CreateThing.html#iot-CreateThing-request-thingName
+ */
+#define MAX_THING_NAME_LENGTH                128
+
+/**
+ * @brief The maximum number of times to run the loop in this demo.
+ *
+ * @note The demo loop is attempted to re-run only if it fails in an iteration.
+ * Once the demo loop succeeds in an iteration, the demo exits successfully.
+ */
+#ifndef FLEET_PROV_MAX_DEMO_LOOP_COUNT
+    #define FLEET_PROV_MAX_DEMO_LOOP_COUNT    ( 3 )
+#endif
+
+/**
+ * @brief Time in seconds to wait between retries of the demo loop if
+ * demo loop fails.
+ */
+#define DELAY_BETWEEN_DEMO_RETRY_ITERATIONS_SECONDS    ( 5 )
+
+/**
+ * @brief Size of buffer in which to hold the certificate signing request (CSR).
+ */
+#define CSR_BUFFER_LENGTH                              2048
+
+/**
+ * @brief Size of buffer in which to hold the certificate.
+ */
+#define CERT_BUFFER_LENGTH                             2048
+
+/**
+ * @brief Size of buffer in which to hold the certificate id.
+ *
+ * See https://docs.aws.amazon.com/iot/latest/apireference/API_Certificate.html#iot-Type-Certificate-certificateId
+ */
+#define CERT_ID_BUFFER_LENGTH                          64
+
+/**
+ * @brief Size of buffer in which to hold the certificate ownership token.
+ */
+#define OWNERSHIP_TOKEN_BUFFER_LENGTH                  512
+
+/**
+ * @brief Status values of the Fleet Provisioning response.
+ */
+typedef enum
+{
+    ResponseNotReceived,
+    ResponseAccepted,
+    ResponseRejected
+} ResponseStatus_t;
+
+/*-----------------------------------------------------------*/
+
+/**
+ * @brief Status reported from the MQTT publish callback.
+ */
+static ResponseStatus_t responseStatus;
+
+/**
+ * @brief Buffer to hold the provisioned AWS IoT Thing name.
+ */
+static char thingName[ MAX_THING_NAME_LENGTH ];
+
+/**
+ * @brief Length of the AWS IoT Thing name.
+ */
+static size_t thingNameLength;
+
+/**
+ * @brief Buffer to hold responses received from the AWS IoT Fleet Provisioning
+ * APIs. When the MQTT publish callback receives an expected Fleet Provisioning
+ * accepted payload, it copies it into this buffer.
+ */
+static uint8_t payloadBuffer[ NETWORK_BUFFER_SIZE ];
+
+/**
+ * @brief Length of the payload stored in #payloadBuffer. This is set by the
+ * MQTT publish callback when it copies a received payload into #payloadBuffer.
+ */
+static size_t payloadLength;
+
+/*-----------------------------------------------------------*/
+
+/**
+ * @brief Callback to receive the incoming publish messages from the MQTT
+ * broker. Sets responseStatus if an expected CreateCertificateFromCsr or
+ * RegisterThing response is received, and copies the response into
+ * responseBuffer if the response is an accepted one.
+ *
+ * @param[in] pPublishInfo Pointer to publish info of the incoming publish.
+ * @param[in] packetIdentifier Packet identifier of the incoming publish.
+ */
+static void provisioningPublishCallback( MQTTPublishInfo_t * pPublishInfo,
+                                         uint16_t packetIdentifier );
+
+/**
+ * @brief Run the MQTT process loop to get a response.
+ */
+static bool waitForResponse( void );
+
+/**
+ * @brief Subscribe to the CreateCertificateFromCsr accepted and rejected topics.
+ */
+static bool subscribeToCsrResponseTopics( void );
+
+/**
+ * @brief Unsubscribe from the CreateCertificateFromCsr accepted and rejected topics.
+ */
+static bool unsubscribeFromCsrResponseTopics( void );
+
+/**
+ * @brief Subscribe to the RegisterThing accepted and rejected topics.
+ */
+static bool subscribeToRegisterThingResponseTopics( void );
+
+/**
+ * @brief Unsubscribe from the RegisterThing accepted and rejected topics.
+ */
+static bool unsubscribeFromRegisterThingResponseTopics( void );
+
+/*-----------------------------------------------------------*/
+
+static void provisioningPublishCallback( MQTTPublishInfo_t * pPublishInfo,
+                                         uint16_t packetIdentifier )
+{
+    FleetProvisioningStatus_t status;
+    FleetProvisioningTopic_t api;
+
+    /* Silence compiler warnings about unused variables. */
+    ( void ) packetIdentifier;
+
+    status = FleetProvisioning_MatchTopic( pPublishInfo->pTopicName,
+                                           pPublishInfo->topicNameLength, &api );
+
+    if( status != FleetProvisioningSuccess )
+    {
+        LogWarn( ( "Unexpected publish message received. Topic: %.*s.",
+                   ( int ) pPublishInfo->topicNameLength,
+                   ( const char * ) pPublishInfo->pTopicName ) );
+    }
+    else
+    {
+        if( api == FleetProvCborCreateCertFromCsrAccepted )
+        {
+            LogInfo( ( "Received accepted response from Fleet Provisioning CreateCertificateFromCsr API." ) );
+
+            responseStatus = ResponseAccepted;
+
+            /* Copy the payload from the MQTT library's buffer to #payloadBuffer. */
+            ( void ) memcpy( ( void * ) payloadBuffer,
+                             ( const void * ) pPublishInfo->pPayload,
+                             ( size_t ) pPublishInfo->payloadLength );
+
+            payloadLength = pPublishInfo->payloadLength;
+        }
+        else if( api == FleetProvCborCreateCertFromCsrRejected )
+        {
+            LogError( ( "Received rejected response from Fleet Provisioning CreateCertificateFromCsr API." ) );
+
+            responseStatus = ResponseRejected;
+        }
+        else if( api == FleetProvCborRegisterThingAccepted )
+        {
+            LogInfo( ( "Received accepted response from Fleet Provisioning RegisterThing API." ) );
+
+            responseStatus = ResponseAccepted;
+
+            /* Copy the payload from the MQTT library's buffer to #payloadBuffer. */
+            ( void ) memcpy( ( void * ) payloadBuffer,
+                             ( const void * ) pPublishInfo->pPayload,
+                             ( size_t ) pPublishInfo->payloadLength );
+
+            payloadLength = pPublishInfo->payloadLength;
+        }
+        else if( api == FleetProvCborCreateCertFromCsrRejected )
+        {
+            LogError( ( "Received rejected response from Fleet Provisioning RegisterThing API." ) );
+
+            responseStatus = ResponseRejected;
+        }
+        else
+        {
+            LogError( ( "Received message on unexpected Fleet Provisioning topic. Topic: %.*s.",
+                        ( int ) pPublishInfo->topicNameLength,
+                        ( const char * ) pPublishInfo->pTopicName ) );
+        }
+    }
+}
+/*-----------------------------------------------------------*/
+
+static bool waitForResponse( void )
+{
+    bool status = false;
+
+    responseStatus = ResponseNotReceived;
+
+    /* responseStatus is updated from the MQTT publish callback. */
+    ( void ) ProcessLoop();
+
+    if( responseStatus == ResponseNotReceived )
+    {
+        LogError( ( "Timed out waiting for response." ) );
+    }
+
+    if( responseStatus == ResponseAccepted )
+    {
+        status = true;
+    }
+
+    return status;
+}
+/*-----------------------------------------------------------*/
+
+static bool subscribeToCsrResponseTopics( void )
+{
+    bool status;
+
+    status = SubscribeToTopic( FP_CBOR_CREATE_CERT_ACCEPTED_TOPIC,
+                               FP_CBOR_CREATE_CERT_ACCEPTED_LENGTH );
+
+    if( status == false )
+    {
+        LogError( ( "Failed to subscribe to fleet provisioning topic: %.*s.",
+                    FP_CBOR_CREATE_CERT_ACCEPTED_LENGTH,
+                    FP_CBOR_CREATE_CERT_ACCEPTED_TOPIC ) );
+    }
+
+    if( status == true )
+    {
+        status = SubscribeToTopic( FP_CBOR_CREATE_CERT_REJECTED_TOPIC,
+                                   FP_CBOR_CREATE_CERT_REJECTED_LENGTH );
+
+        if( status == false )
+        {
+            LogError( ( "Failed to subscribe to fleet provisioning topic: %.*s.",
+                        FP_CBOR_CREATE_CERT_REJECTED_LENGTH,
+                        FP_CBOR_CREATE_CERT_REJECTED_TOPIC ) );
+        }
+    }
+
+    return status;
+}
+/*-----------------------------------------------------------*/
+
+static bool unsubscribeFromCsrResponseTopics( void )
+{
+    bool status;
+
+    status = UnsubscribeFromTopic( FP_CBOR_CREATE_CERT_ACCEPTED_TOPIC,
+                                   FP_CBOR_CREATE_CERT_ACCEPTED_LENGTH );
+
+    if( status == false )
+    {
+        LogError( ( "Failed to unsubscribe from fleet provisioning topic: %.*s.",
+                    FP_CBOR_CREATE_CERT_ACCEPTED_LENGTH,
+                    FP_CBOR_CREATE_CERT_ACCEPTED_TOPIC ) );
+    }
+
+    if( status == true )
+    {
+        status = UnsubscribeFromTopic( FP_CBOR_CREATE_CERT_REJECTED_TOPIC,
+                                       FP_CBOR_CREATE_CERT_REJECTED_LENGTH );
+
+        if( status == false )
+        {
+            LogError( ( "Failed to unsubscribe from fleet provisioning topic: %.*s.",
+                        FP_CBOR_CREATE_CERT_REJECTED_LENGTH,
+                        FP_CBOR_CREATE_CERT_REJECTED_TOPIC ) );
+        }
+    }
+
+    return status;
+}
+/*-----------------------------------------------------------*/
+
+static bool subscribeToRegisterThingResponseTopics( void )
+{
+    bool status;
+
+    status = SubscribeToTopic( FP_CBOR_REGISTER_ACCEPTED_TOPIC( PROVISIONING_TEMPLATE_NAME ),
+                               FP_CBOR_REGISTER_ACCEPTED_LENGTH( PROVISIONING_TEMPLATE_NAME_LENGTH ) );
+
+    if( status == false )
+    {
+        LogError( ( "Failed to subscribe to fleet provisioning topic: %.*s.",
+                    FP_CBOR_REGISTER_ACCEPTED_LENGTH( PROVISIONING_TEMPLATE_NAME_LENGTH ),
+                    FP_CBOR_REGISTER_ACCEPTED_TOPIC( PROVISIONING_TEMPLATE_NAME ) ) );
+    }
+
+    if( status == true )
+    {
+        status = SubscribeToTopic( FP_CBOR_REGISTER_REJECTED_TOPIC( PROVISIONING_TEMPLATE_NAME ),
+                                   FP_CBOR_REGISTER_REJECTED_LENGTH( PROVISIONING_TEMPLATE_NAME_LENGTH ) );
+
+        if( status == false )
+        {
+            LogError( ( "Failed to subscribe to fleet provisioning topic: %.*s.",
+                        FP_CBOR_REGISTER_REJECTED_LENGTH( PROVISIONING_TEMPLATE_NAME_LENGTH ),
+                        FP_CBOR_REGISTER_REJECTED_TOPIC( PROVISIONING_TEMPLATE_NAME ) ) );
+        }
+    }
+
+    return status;
+}
+/*-----------------------------------------------------------*/
+
+static bool unsubscribeFromRegisterThingResponseTopics( void )
+{
+    bool status;
+
+    status = UnsubscribeFromTopic( FP_CBOR_REGISTER_ACCEPTED_TOPIC( PROVISIONING_TEMPLATE_NAME ),
+                                   FP_CBOR_REGISTER_ACCEPTED_LENGTH( PROVISIONING_TEMPLATE_NAME_LENGTH ) );
+
+    if( status == false )
+    {
+        LogError( ( "Failed to unsubscribe from fleet provisioning topic: %.*s.",
+                    FP_CBOR_REGISTER_ACCEPTED_LENGTH( PROVISIONING_TEMPLATE_NAME_LENGTH ),
+                    FP_CBOR_REGISTER_ACCEPTED_TOPIC( PROVISIONING_TEMPLATE_NAME ) ) );
+    }
+
+    if( status == true )
+    {
+        status = UnsubscribeFromTopic( FP_CBOR_REGISTER_REJECTED_TOPIC( PROVISIONING_TEMPLATE_NAME ),
+                                       FP_CBOR_REGISTER_REJECTED_LENGTH( PROVISIONING_TEMPLATE_NAME_LENGTH ) );
+
+        if( status == false )
+        {
+            LogError( ( "Failed to unsubscribe from fleet provisioning topic: %.*s.",
+                        FP_CBOR_REGISTER_REJECTED_LENGTH( PROVISIONING_TEMPLATE_NAME_LENGTH ),
+                        FP_CBOR_REGISTER_REJECTED_TOPIC( PROVISIONING_TEMPLATE_NAME ) ) );
+        }
+    }
+
+    return status;
+}
+/*-----------------------------------------------------------*/
+
+/* This example uses a single application task, which shows that how to use
+ * the Fleet Provisioning library to generate and validate AWS IoT Fleet
+ * Provisioning MQTT topics, and use the coreMQTT library to communicate with
+ * the AWS IoT Fleet Provisioning APIs. */
+int vStartFleetProvisioningDemo()
+{
+    bool status = false;
+    /* Buffer for holding the CSR. */
+    char csr[ CSR_BUFFER_LENGTH ] = { 0 };
+    size_t csrLength = 0;
+    /* Buffer for holding received certificate until it is saved. */
+    char certificate[ CERT_BUFFER_LENGTH ];
+    size_t certificateLength;
+    /* Buffer for holding the certificate ID. */
+    char certificateId[ CERT_ID_BUFFER_LENGTH ];
+    size_t certificateIdLength;
+    /* Buffer for holding the certificate ownership token. */
+    char ownershipToken[ OWNERSHIP_TOKEN_BUFFER_LENGTH ];
+    size_t ownershipTokenLength;
+    bool connectionEstablished = false;
+    CK_SESSION_HANDLE p11Session;
+    int demoRunCount = 0;
+    CK_RV pkcs11ret = CKR_OK;
+
+    do
+    {
+        /* Initialize the buffer lengths to their max lengths. */
+        certificateLength = CERT_BUFFER_LENGTH;
+        certificateIdLength = CERT_ID_BUFFER_LENGTH;
+        ownershipTokenLength = OWNERSHIP_TOKEN_BUFFER_LENGTH;
+
+        /* Initialize the PKCS #11 module */
+        pkcs11ret = xInitializePkcs11Session( &p11Session );
+
+        if( pkcs11ret != CKR_OK )
+        {
+            LogError( ( "Failed to initialize PKCS #11." ) );
+            status = false;
+        }
+        else
+        {
+            /* Insert the claim credentials into the PKCS #11 module */
+            status = loadClaimCredentials( p11Session,
+                                           CLAIM_CERT_PATH,
+                                           pkcs11configLABEL_CLAIM_CERTIFICATE,
+                                           CLAIM_PRIVATE_KEY_PATH,
+                                           pkcs11configLABEL_CLAIM_PRIVATE_KEY );
+
+            if( status == false )
+            {
+                LogError( ( "Failed to provision PKCS #11 with claim credentials." ) );
+            }
+        }
+
+        /**** Connect to AWS IoT Core with provisioning claim credentials *****/
+
+        /* We first use the claim credentials to connect to the broker. These
+         * credentials should allow use of the RegisterThing API and one of the
+         * CreateCertificatefromCsr or CreateKeysAndCertificate.
+         * In this demo we use CreateCertificatefromCsr. */
+
+        if( status == true )
+        {
+            /* Attempts to connect to the AWS IoT MQTT broker. If the
+             * connection fails, retries after a timeout. Timeout value will
+             * exponentially increase until maximum attempts are reached. */
+            LogInfo( ( "Establishing MQTT session with claim certificate..." ) );
+            status = EstablishMqttSession( provisioningPublishCallback,
+                                           p11Session,
+                                           pkcs11configLABEL_CLAIM_CERTIFICATE,
+                                           pkcs11configLABEL_CLAIM_PRIVATE_KEY );
+
+            if( status == false )
+            {
+                LogError( ( "Failed to establish MQTT session." ) );
+            }
+            else
+            {
+                LogInfo( ( "Established connection with claim credentials." ) );
+                connectionEstablished = true;
+            }
+        }
+
+        /**** Call the CreateCertificateFromCsr API ***************************/
+
+        /* We use the CreateCertificatefromCsr API to obtain a client certificate
+         * for a key on the device by means of sending a certificate signing
+         * request (CSR). */
+        if( status == true )
+        {
+            /* Subscribe to the CreateCertificateFromCsr accepted and rejected
+             * topics. In this demo we use CBOR encoding for the payloads,
+             * so we use the CBOR variants of the topics. */
+            status = subscribeToCsrResponseTopics();
+        }
+
+        if( status == true )
+        {
+            /* Create a new key and CSR. */
+            status = generateKeyAndCsr( p11Session,
+                                        pkcs11configLABEL_DEVICE_PRIVATE_KEY_FOR_TLS,
+                                        pkcs11configLABEL_DEVICE_PUBLIC_KEY_FOR_TLS,
+                                        csr,
+                                        CSR_BUFFER_LENGTH,
+                                        &csrLength );
+        }
+
+        if( status == true )
+        {
+            /* Create the request payload containing the CSR to publish to the
+             * CreateCertificateFromCsr APIs. */
+            status = generateCsrRequest( payloadBuffer,
+                                         NETWORK_BUFFER_SIZE,
+                                         csr,
+                                         csrLength,
+                                         &payloadLength );
+        }
+
+        if( status == true )
+        {
+            /* Publish the CSR to the CreateCertificatefromCsr API. */
+            PublishToTopic( FP_CBOR_CREATE_CERT_PUBLISH_TOPIC,
+                            FP_CBOR_CREATE_CERT_PUBLISH_LENGTH,
+                            ( char * ) payloadBuffer,
+                            payloadLength );
+
+            if( status == false )
+            {
+                LogError( ( "Failed to publish to fleet provisioning topic: %.*s.",
+                            FP_CBOR_CREATE_CERT_PUBLISH_LENGTH,
+                            FP_CBOR_CREATE_CERT_PUBLISH_TOPIC ) );
+            }
+        }
+
+        if( status == true )
+        {
+            /* Get the response to the CreateCertificatefromCsr request. */
+            status = waitForResponse();
+        }
+
+        if( status == true )
+        {
+            /* From the response, extract the certificate, certificate ID, and
+             * certificate ownership token. */
+            status = parseCsrResponse( payloadBuffer,
+                                       payloadLength,
+                                       certificate,
+                                       &certificateLength,
+                                       certificateId,
+                                       &certificateIdLength,
+                                       ownershipToken,
+                                       &ownershipTokenLength );
+
+            if( status == true )
+            {
+                LogInfo( ( "Received certificate with Id: %.*s", ( int ) certificateIdLength, certificateId ) );
+            }
+        }
+
+        if( status == true )
+        {
+            /* Save the certificate into PKCS #11. */
+            status = loadCertificate( p11Session,
+                                      certificate,
+                                      pkcs11configLABEL_DEVICE_CERTIFICATE_FOR_TLS,
+                                      certificateLength );
+        }
+
+        if( status == true )
+        {
+            /* Unsubscribe from the CreateCertificateFromCsr topics. */
+            status = unsubscribeFromCsrResponseTopics();
+        }
+
+        /**** Call the RegisterThing API **************************************/
+
+        /* We then use the RegisterThing API to activate the received certificate,
+         * provision AWS IoT resources according to the provisioning template, and
+         * receive device configuration. */
+        if( status == true )
+        {
+            /* Create the request payload to publish to the RegisterThing API. */
+            status = generateRegisterThingRequest( payloadBuffer,
+                                                   NETWORK_BUFFER_SIZE,
+                                                   ownershipToken,
+                                                   ownershipTokenLength,
+                                                   DEVICE_SERIAL_NUMBER,
+                                                   DEVICE_SERIAL_NUMBER_LENGTH,
+                                                   &payloadLength );
+        }
+
+        if( status == true )
+        {
+            /* Subscribe to the RegisterThing response topics. */
+            status = subscribeToRegisterThingResponseTopics();
+        }
+
+        if( status == true )
+        {
+            /* Publish the RegisterThing request. */
+            PublishToTopic( FP_CBOR_REGISTER_PUBLISH_TOPIC( PROVISIONING_TEMPLATE_NAME ),
+                            FP_CBOR_REGISTER_PUBLISH_LENGTH( PROVISIONING_TEMPLATE_NAME_LENGTH ),
+                            ( char * ) payloadBuffer,
+                            payloadLength );
+
+            if( status == false )
+            {
+                LogError( ( "Failed to publish to fleet provisioning topic: %.*s.",
+                            FP_CBOR_REGISTER_PUBLISH_LENGTH( PROVISIONING_TEMPLATE_NAME_LENGTH ),
+                            FP_CBOR_REGISTER_PUBLISH_TOPIC( PROVISIONING_TEMPLATE_NAME ) ) );
+            }
+        }
+
+        if( status == true )
+        {
+            /* Get the response to the RegisterThing request. */
+            status = waitForResponse();
+        }
+
+        if( status == true )
+        {
+            /* Extract the Thing name from the response. */
+            thingNameLength = MAX_THING_NAME_LENGTH;
+            status = parseRegisterThingResponse( payloadBuffer,
+                                                 payloadLength,
+                                                 thingName,
+                                                 &thingNameLength );
+
+            if( status == true )
+            {
+                LogInfo( ( "Received AWS IoT Thing name: %.*s", ( int ) thingNameLength, thingName ) );
+            }
+        }
+
+        if( status == true )
+        {
+            /* Unsubscribe from the RegisterThing topics. */
+            unsubscribeFromRegisterThingResponseTopics();
+        }
+
+        /**** Disconnect from AWS IoT Core ************************************/
+
+        /* As we have completed the provisioning workflow, we disconnect from
+         * the connection using the provisioning claim credentials. We will
+         * establish a new MQTT connection with the newly provisioned
+         * credentials. */
+        if( connectionEstablished == true )
+        {
+            DisconnectMqttSession();
+            connectionEstablished = false;
+        }
+
+        /**** Connect to AWS IoT Core with provisioned certificate ************/
+
+        if( status == true )
+        {
+            LogInfo( ( "Establishing MQTT session with provisioned certificate..." ) );
+            status = EstablishMqttSession( provisioningPublishCallback,
+                                           p11Session,
+                                           pkcs11configLABEL_DEVICE_CERTIFICATE_FOR_TLS,
+                                           pkcs11configLABEL_DEVICE_PRIVATE_KEY_FOR_TLS );
+
+            if( status != true )
+            {
+                LogError( ( "Failed to establish MQTT session with provisioned "
+                            "credentials. Verify on your AWS account that the "
+                            "new certificate is active and has an attached IoT "
+                            "Policy that allows the \"iot:Connect\" action." ) );
+            }
+            else
+            {
+                LogInfo( ( "Sucessfully established connection with provisioned credentials." ) );
+                connectionEstablished = true;
+            }
+        }
+
+        /**** Finish **********************************************************/
+
+        if( connectionEstablished == true )
+        {
+            /* Close the connection. */
+            DisconnectMqttSession();
+            connectionEstablished = false;
+        }
+
+        pkcs11CloseSession( p11Session );
+
+        /**** Retry in case of failure ****************************************/
+
+        /* Increment the demo run count. */
+        demoRunCount++;
+
+        if( status == true )
+        {
+            LogInfo( ( "Demo iteration %d is successful.", demoRunCount ) );
+        }
+        /* Attempt to retry a failed iteration of demo for up to #FLEET_PROV_MAX_DEMO_LOOP_COUNT times. */
+        else if( demoRunCount < FLEET_PROV_MAX_DEMO_LOOP_COUNT )
+        {
+            LogWarn( ( "Demo iteration %d failed. Retrying...", demoRunCount ) );
+            sleep( DELAY_BETWEEN_DEMO_RETRY_ITERATIONS_SECONDS );
+        }
+        /* Failed all #FLEET_PROV_MAX_DEMO_LOOP_COUNT demo iterations. */
+        else
+        {
+            LogError( ( "All %d demo iterations failed.", FLEET_PROV_MAX_DEMO_LOOP_COUNT ) );
+            break;
+        }
+    } while( status != true );
+
+    /* Log demo success. */
+    if( status == true )
+    {
+        LogInfo( ( "Demo completed successfully." ) );
+    }
+
+    return ( status == true ) ? EXIT_SUCCESS : EXIT_FAILURE;
+}
+/*-----------------------------------------------------------*/
